makefileok

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gdel-cas <gdel-cas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 15:26:46 by gdel-cas          #+#    #+#             */
-/*   Updated: 2023/09/18 16:59:45 by gdel-cas         ###   ########.fr       */
+/*   Updated: 2023/09/18 18:29:15 by gdel-cas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <stdio.h>
@@ -15,7 +15,7 @@
 
 void    *ft_memset(void *b, int c, size_t n)
 {
-    int i;
+    size_t	i;
     unsigned char *ptr;
     
     i = 0;
